refactor: WIP tidy up ColumnDisplayInfo
